Add ShellUtils to define shell related utils since they don't belong in BackgroundJob
